feat: Add buttons for refreshing contracts and accounts
This commit adds buttons for refreshing contracts and accounts in the WasmVmPanel. The buttons are connected to the "refreshContractsBtn" and "refreshAccBtn" elements respectively. When clicked, these buttons trigger the "unimplemented" function, which sends a message to the vscode extension to handle the refresh action. This feature was added to improve the user experience by allowing them to easily refresh the contracts and accounts data in the panel.